+++ src/evmla/ethereal_ir/function/block/element/mod.rs
@@ -724,7 +724,7 @@ where
                 );
 
                 context.build_memcpy(
-                    context.intrinsics().memory_copy,
+                    context.intrinsics().memory_move,
                     destination,
                     source,
                     arguments[2].into_int_value(),
@@ -1132,6 +1132,7 @@ where
 
                 era_compiler_llvm_context::eravm_evm_create::create(
                     context,
+                    era_compiler_llvm_context::EraVMAddressSpace::Heap,
                     value,
                     input_offset,
                     input_length,
@@ -1148,6 +1149,7 @@ where
 
                 era_compiler_llvm_context::eravm_evm_create::create2(
                     context,
+                    era_compiler_llvm_context::EraVMAddressSpace::Heap,
                     value,
                     input_offset,
                     input_length,