+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # The `zksolc` changelog
 
-## [Unreleased]
+## [1.4.0] - 2024-02-19
 
 ### Added
 
@@ -8,6 +8,11 @@
 - The `MCOPY` instruction support
 - The `--evm-version` parameter to the CLI
 - An option to disable the system request memoization
+- More compiler optimizations
+
+### Fixed
+
+- An issue with `MCOPY` for overlapping memory regions
 
 ## [1.3.23] - 2024-01-30